refactor(examples): centralize asset-bytes conversions
Pull every conversion between image bytes, data URLs, and R2 object
storage into one asset-bytes module: blobToDataUrl (now covering the
former file/blob readers in the sidebar and brand panel), urlToDataUrl,
and uploadImageBytes. multiplayerAssetStore is left as just the
TLAssetStore adapter; the programmatic helpers no longer live alongside
it. Concentrates ADR-0006's upload endpoint, object naming, and
read-back in a single place.
